RELEASE = 'php artisan list' version: 2.0.3

+++ database/migrations/2021_11_15_114344_create_static_page_blocks_table.php
@@ -15,7 +15,7 @@ class CreateStaticPageBlocksTable extends Migration
     {
         Schema::create('static_page_blocks', function (Blueprint $table) {
             $table->id();
-            $table->unsignedInteger('static_page_type_id');
+            $table->unsignedSmallInteger('static_page_type_id');
             $table->unsignedTinyInteger('show_in')
                 ->comment('1=>Nise3, 2=> Youth, 3=>TSP, 4=>Industry, 5=>Industry Association');
 
@@ -46,7 +46,7 @@ class CreateStaticPageBlocksTable extends Migration
             $table->unsignedInteger('created_by')->nullable();
             $table->unsignedInteger('updated_by')->nullable();
 
-            $table->tinyInteger('row_status')
+            $table->unsignedTinyInteger('row_status')
                 ->default(1)
                 ->comment('ACTIVE_STATUS => 1, INACTIVE_STATUS => 0');
